django/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
attrs==25.1.0
Django==5.1.6
djangorestframework==3.15.2
idna==3.10
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
requests==2.32.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
urllib3==2.3.0
typing_extensions==4.12.2
uritemplate==4.1.1
